+++ benches/e2e_benchmark.rs
@@ -3,9 +3,12 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use rand::{prelude::IteratorRandom, rngs::OsRng, CryptoRng, Rng, RngCore};
 use std::collections::HashMap;
 use tss_ecdsa::{
-    auxinfo::AuxInfoParticipant, errors::Result, keygen::KeygenParticipant, messages::Message,
-    Identifier, Participant, ParticipantConfig, ParticipantIdentifier, PresignInput,
-    PresignParticipant, ProtocolParticipant,
+    auxinfo::AuxInfoParticipant,
+    errors::Result,
+    keygen::KeygenParticipant,
+    messages::Message,
+    presign::{self, PresignParticipant},
+    Identifier, Participant, ParticipantConfig, ParticipantIdentifier, ProtocolParticipant,
 };
 
 /// Delivers all messages into their respective participant's inboxes
@@ -134,7 +137,7 @@ fn run_benchmarks_for_given_size(c: &mut Criterion, num_players: usize) {
         .into_iter()
         .zip(keygen_outputs)
         .map(|(auxinfo_output, keygen_output)| {
-            PresignInput::new(auxinfo_output, keygen_output).unwrap()
+            presign::Input::new(auxinfo_output, keygen_output).unwrap()
         })
         .collect::<Vec<_>>();
